fix(sui/coins): store timestamp as BigInt to avoid SDK 4 int32 overflow
The Metadata.timestamp store field was `Int` (proto RichValue.int_value = int32) but holds an epoch-ms value (ctx.timestamp.getTime() ~1.68e12), which overflows int32. SDK 3's ts-proto didn't range-check int32 on encode so it passed silently; SDK 4's protobuf-es validates strictly and throws at serialization in the onObjectChange handler. Change the field to BigInt (64-bit) and pass BigInt(...) at both upsert sites.

+++ sui/coins/src/schema/store.ts
@@ -22,7 +22,7 @@ interface MetadataConstructorInput {
   description: String;
   icon_url?: String;
   digest: String;
-  timestamp: Int;
+  timestamp: BigInt;
 }
 @Entity("Metadata")
 export class Metadata extends AbstractEntity  {
@@ -59,8 +59,8 @@ export class Metadata extends AbstractEntity  {
 	digest: String
 
 	@Required
-	@Column("Int")
-	timestamp: Int
+	@Column("BigInt")
+	timestamp: BigInt
   constructor(data: Partial<MetadataConstructorInput>) {super()}
   
   
@@ -79,7 +79,7 @@ const source = `type Metadata @entity {
     description: String! # Description of the token
     icon_url: String
     digest: String! # transaction digest created the metadata
-    timestamp: Int! # timestamp of the metadata creation
+    timestamp: BigInt! # timestamp of the metadata creation (epoch ms; BigInt to avoid int32 overflow)
 }`
 DatabaseSchema.register({
   source,
